Additional Files to corrected to bigint

+++ administrator/components/com_finder/sql/install.mysql.sql
@@ -3,18 +3,18 @@
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_filters` (
-  `filter_id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `filter_id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
   `title` varchar(255) NOT NULL,
   `alias` varchar(255) NOT NULL,
   `state` tinyint(1) NOT NULL DEFAULT 1,
   `created` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
-  `created_by` int(10) unsigned NOT NULL,
+  `created_by` bigint(10) unsigned NOT NULL,
   `created_by_alias` varchar(255) NOT NULL,
   `modified` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
-  `modified_by` int(10) unsigned NOT NULL DEFAULT 0,
-  `checked_out` int(10) unsigned NOT NULL DEFAULT 0,
+  `modified_by` bigint(10) unsigned NOT NULL DEFAULT 0,
+  `checked_out` bigint(10) unsigned NOT NULL DEFAULT 0,
   `checked_out_time` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
-  `map_count` int(10) unsigned NOT NULL DEFAULT 0,
+  `map_count` bigint(10) unsigned NOT NULL DEFAULT 0,
   `data` text NOT NULL,
   `params` mediumtext,
   PRIMARY KEY (`filter_id`)
@@ -25,7 +25,7 @@ CREATE TABLE IF NOT EXISTS `#__finder_filters` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links` (
-  `link_id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `link_id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
   `url` varchar(255) NOT NULL,
   `route` varchar(255) NOT NULL,
   `title` varchar(400) DEFAULT NULL,
@@ -34,7 +34,7 @@ CREATE TABLE IF NOT EXISTS `#__finder_links` (
   `md5sum` varchar(32) DEFAULT NULL,
   `published` tinyint(1) NOT NULL DEFAULT 1,
   `state` int(5) DEFAULT 1,
-  `access` int(5) DEFAULT 0,
+  `access` bigint(10) unsigned DEFAULT 0,
   `language` varchar(8) NOT NULL,
   `publish_start_date` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
   `publish_end_date` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS `#__finder_links` (
   `end_date` datetime NOT NULL DEFAULT '0000-00-00 00:00:00',
   `list_price` double unsigned NOT NULL DEFAULT 0,
   `sale_price` double unsigned NOT NULL DEFAULT 0,
-  `type_id` int(11) NOT NULL,
+  `type_id` bigint(10) NOT NULL,
   `object` mediumblob NOT NULL,
   PRIMARY KEY (`link_id`),
   KEY `idx_type` (`type_id`),
@@ -58,8 +58,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms0` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -71,8 +71,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms0` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms1` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -84,8 +84,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms1` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms2` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -97,8 +97,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms2` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms3` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -110,8 +110,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms3` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms4` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -123,8 +123,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms4` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms5` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -136,8 +136,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms5` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms6` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -149,8 +149,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms6` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms7` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -162,8 +162,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms7` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms8` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -175,8 +175,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms8` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_terms9` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -188,8 +188,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_terms9` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_termsa` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -201,8 +201,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_termsa` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_termsb` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -214,8 +214,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_termsb` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_termsc` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -227,8 +227,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_termsc` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_termsd` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -240,8 +240,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_termsd` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_termse` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -253,8 +253,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_termse` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_links_termsf` (
-  `link_id` int(10) unsigned NOT NULL,
-  `term_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `weight` float unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`term_id`),
   KEY `idx_term_weight` (`term_id`,`weight`),
@@ -266,8 +266,8 @@ CREATE TABLE IF NOT EXISTS `#__finder_links_termsf` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_taxonomy` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-  `parent_id` int(10) unsigned NOT NULL DEFAULT 0,
+  `id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
+  `parent_id` bigint(10) unsigned NOT NULL DEFAULT 0,
   `title` varchar(255) NOT NULL,
   `state` tinyint(1) unsigned NOT NULL DEFAULT 1,
   `access` tinyint(1) unsigned NOT NULL DEFAULT 0,
@@ -292,8 +292,8 @@ REPLACE INTO `#__finder_taxonomy` (`id`, `parent_id`, `title`, `state`, `access`
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_taxonomy_map` (
-  `link_id` int(10) unsigned NOT NULL,
-  `node_id` int(10) unsigned NOT NULL,
+  `link_id` bigint(10) unsigned NOT NULL,
+  `node_id` bigint(10) unsigned NOT NULL,
   PRIMARY KEY (`link_id`,`node_id`),
   KEY `link_id` (`link_id`),
   KEY `node_id` (`node_id`)
@@ -304,14 +304,14 @@ CREATE TABLE IF NOT EXISTS `#__finder_taxonomy_map` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_terms` (
-  `term_id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `term_id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
   `term` varchar(75) NOT NULL,
   `stem` varchar(75) NOT NULL,
   `common` tinyint(1) unsigned NOT NULL DEFAULT 0,
   `phrase` tinyint(1) unsigned NOT NULL DEFAULT 0,
   `weight` float unsigned NOT NULL DEFAULT 0,
   `soundex` varchar(75) NOT NULL,
-  `links` int(10) NOT NULL DEFAULT 0,
+  `links` bigint(10) NOT NULL DEFAULT 0,
   `language` char(3) NOT NULL DEFAULT '',
   PRIMARY KEY (`term_id`),
   UNIQUE KEY `idx_term` (`term`),
@@ -473,7 +473,7 @@ CREATE TABLE IF NOT EXISTS `#__finder_tokens` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_tokens_aggregate` (
-  `term_id` int(10) unsigned NOT NULL,
+  `term_id` bigint(10) unsigned NOT NULL,
   `map_suffix` char(1) NOT NULL,
   `term` varchar(75) NOT NULL,
   `stem` varchar(75) NOT NULL,
@@ -493,7 +493,7 @@ CREATE TABLE IF NOT EXISTS `#__finder_tokens_aggregate` (
 --
 
 CREATE TABLE IF NOT EXISTS `#__finder_types` (
-  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `id` bigint(10) unsigned NOT NULL AUTO_INCREMENT,
   `title` varchar(100) NOT NULL,
   `mime` varchar(100) NOT NULL,
   PRIMARY KEY (`id`),
